Bump to v1.0.1

+++ maia/patches/v1_0/drop_patient_column.py
@@ -6,6 +6,11 @@ def execute():
     Drop patient column in Contact, Customer and Sales Invoice Tables
     """
 
-    frappe.db.sql("""ALTER TABLE `tabSales Invoice` DROP COLUMN patient""")
-    frappe.db.sql("""ALTER TABLE `tabCustomer` DROP COLUMN patient""")
-    frappe.db.sql("""ALTER TABLE `tabContact` DROP COLUMN patient""")
+    if "patient" in frappe.db.get_table_columns("Sales Invoice"):
+        frappe.db.sql("""ALTER TABLE `tabSales Invoice` DROP COLUMN patient""")
+
+    if "patient" in frappe.db.get_table_columns("Customer"):
+        frappe.db.sql("""ALTER TABLE `tabCustomer` DROP COLUMN patient""")
+
+    if "patient" in frappe.db.get_table_columns("Contact"):
+        frappe.db.sql("""ALTER TABLE `tabContact` DROP COLUMN patient""")
